requirements.txt
playwright
playwright-stealth
google-genai
pillow
pillow-heif
pandas